+++ Editor/ValidationSuite/TextReport.cs
@@ -17,9 +17,13 @@ namespace UnityEditor.PackageManager.ValidationSuite
         internal void Initialize(VettingContext context)
         {
             var packageInfo = context.ProjectPackageInfo;
-            Write(string.Format(
-                "Validation Suite Results for package \"{0}\"\r\n - Path: {1}\r\n - Version: {2}\r\n - Test Time: {3}\r\n - Tested with {4} version: {5}\r\n", 
-                packageInfo.name, packageInfo.path, packageInfo.version, DateTime.Now, context.VSuiteInfo.name, context.VSuiteInfo.version)
+            Write(
+                string.Format("Validation Suite Results for package \"{0}\"\r\n", packageInfo.name) +
+                string.Format(" - Path: {0}\r\n", packageInfo.path) +
+                string.Format(" - Version: {0}\r\n", packageInfo.version) +
+                string.Format(" - Lifecycle: {0}\r\n", packageInfo.lifecycle) +
+                string.Format(" - Test Time: {0}\r\n", DateTime.Now) +
+                string.Format(" - Tested with {0} version: {1}\r\n", context.VSuiteInfo.name, context.VSuiteInfo.version)
             );
 
             if (context.ProjectPackageInfo.dependencies.Any())